fix/be: 회원가입 기능 및 Security 요청사항 수정

+++ backend/src/main/java/com/back/domain/member/service/MemberService.java
@@ -49,7 +49,7 @@ public class MemberService {
         return memberRepository.findByRefreshToken(refreshToken);
     }
 
-    public Optional<Member> findById(int id) {
+    public Optional<Member> findById(long id) {
         return memberRepository.findById(id);
     }
